+++ examples/server.cc
@@ -876,7 +876,7 @@ int recv_stream_data(ngtcp2_conn *conn, int64_t stream_id, int fin,
 
 namespace {
 int acked_crypto_offset(ngtcp2_conn *conn, ngtcp2_crypto_level crypto_level,
-                        uint64_t offset, size_t datalen, void *user_data) {
+                        uint64_t offset, uint64_t datalen, void *user_data) {
   auto h = static_cast<Handler *>(user_data);
   h->remove_tx_crypto_data(crypto_level, offset, datalen);
   return 0;
@@ -885,7 +885,7 @@ int acked_crypto_offset(ngtcp2_conn *conn, ngtcp2_crypto_level crypto_level,
 
 namespace {
 int acked_stream_data_offset(ngtcp2_conn *conn, int64_t stream_id,
-                             uint64_t offset, size_t datalen, void *user_data,
+                             uint64_t offset, uint64_t datalen, void *user_data,
                              void *stream_user_data) {
   auto h = static_cast<Handler *>(user_data);
   if (h->acked_stream_data_offset(stream_id, datalen) != 0) {
@@ -895,7 +895,7 @@ int acked_stream_data_offset(ngtcp2_conn *conn, int64_t stream_id,
 }
 } // namespace
 
-int Handler::acked_stream_data_offset(int64_t stream_id, size_t datalen) {
+int Handler::acked_stream_data_offset(int64_t stream_id, uint64_t datalen) {
   if (!httpconn_) {
     return 0;
   }
@@ -2070,7 +2070,7 @@ void remove_tx_stream_data(std::deque<Buffer> &d, uint64_t &tx_offset,
 } // namespace
 
 void Handler::remove_tx_crypto_data(ngtcp2_crypto_level crypto_level,
-                                    uint64_t offset, size_t datalen) {
+                                    uint64_t offset, uint64_t datalen) {
   auto &crypto = crypto_[crypto_level];
   ::remove_tx_stream_data(crypto.data, crypto.acked_offset, offset + datalen);
 }
